k8s: move filterPodLabels to k8s/utils package for SanitizePodLabels
Currently GetPodMetadata is the only caller of SanitizePodLabels but
other callers will be introduced in successive changes. This change
ensures the io.cilium.k8s.* labels are filtered for these callers as
well.
